Part 24 & 25 - Authentication & Authorization

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
 use App\Http\Controllers\ExtracurricularController;
 use App\Http\Controllers\StudentController;
@@ -18,29 +19,31 @@ use Illuminate\Support\Facades\Route;
  */
 
 Route::get('/', function () {
-    return view('home', [
-        'title' => 'Home',
-        'name' => 'Laizy',
-        'role' => 'admin',
-        'buah' => ['apel', 'pisang', 'kiwi', 'jeruk']]);
-});
+    return view('home');
+})->middleware('auth');
 
-Route::get('/students', [StudentController::class, 'index']);
-Route::get('/students/{id}', [StudentController::class, 'show']);
-Route::get('/students-add', [StudentController::class, 'create']);
-Route::post('/student', [StudentController::class, 'store']);
-Route::get('/students-edit/{id}', [StudentController::class, 'edit']);
-Route::put('/student/{id}', [StudentController::class, 'update']);
-Route::get('/students-delete/{id}', [StudentController::class, 'delete']);
-Route::delete('/students-destroy/{id}', [StudentController::class, 'destroy']);
-Route::get('/students-deleted', [StudentController::class, 'deletedStudent']);
-Route::get('/students-restore/{id}', [StudentController::class, 'restore']);
+Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/login', [AuthController::class, 'auth'])->middleware('guest');
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
 
-Route::get('/class', [ClassController::class, 'index']);
-Route::get('/class/{id}', [ClassController::class, 'show']);
+Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
 
-Route::get('/extracurricular', [ExtracurricularController::class, 'index']);
-Route::get('/extracurricular/{id}', [ExtracurricularController::class, 'show']);
+Route::get('/students/{id}', [StudentController::class, 'show'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/students-add', [StudentController::class, 'create'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::post('/student', [StudentController::class, 'store'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/students-edit/{id}', [StudentController::class, 'edit'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::put('/student/{id}', [StudentController::class, 'update'])->middleware(['auth', 'must-admin-or-teacher']);
 
-Route::get('/teachers', [TeacherController::class, 'index']);
-Route::get('/teachers/{id}', [TeacherController::class, 'show']);
+Route::get('/students-delete/{id}', [StudentController::class, 'delete'])->middleware(['auth', 'must-admin']);
+Route::delete('/students-destroy/{id}', [StudentController::class, 'destroy'])->middleware(['auth', 'must-admin']);
+Route::get('/students-deleted', [StudentController::class, 'deletedStudent'])->middleware(['auth', 'must-admin']);
+Route::get('/students-restore/{id}', [StudentController::class, 'restore'])->middleware(['auth', 'must-admin']);
+
+Route::get('/class', [ClassController::class, 'index'])->middleware('auth');
+Route::get('/class/{id}', [ClassController::class, 'show'])->middleware('auth');
+
+Route::get('/extracurricular', [ExtracurricularController::class, 'index'])->middleware('auth');
+Route::get('/extracurricular/{id}', [ExtracurricularController::class, 'show'])->middleware('auth');
+
+Route::get('/teachers', [TeacherController::class, 'index'])->middleware('auth');
+Route::get('/teachers/{id}', [TeacherController::class, 'show'])->middleware('auth');
